requirements_discussion.txt
python-telegram-bot[job-queue]==20.7
google-generativeai
